kafka_realtime_pipeline/requirements.txt
kafka-python==2.0.2
faker==19.3.1
psycopg2-binary==2.9.7
pandas==2.0.3
plotly==5.16.1
streamlit==1.26.0
sqlalchemy==2.0.20

# Apache Flink PyFlink (BONUS)
apache-flink==1.17.1

# Machine Learning libraries (BONUS)
scikit-learn==1.3.0
numpy==1.24.3
xgboost==1.7.6
imbalanced-learn==0.11.0

# Additional utilities
pyarrow==12.0.1
